Added errors to Bounded
Now you can know if the operation exceeded the bounds and then break - rather than having to explicitly check beforehand.  It's up to the caller to decide how to do that.

I agree this breaks fluent design, but it's REALLY crucial to everything else.  The other types don't have to propagate the error - only handle it if they care